# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorDTO:
    properties:
      error:
        example: Something went wrong
        type: string
    type: object
...
  dto.ProfileResponseDTO:
    properties:
      avatar_file_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      bio:
        example: Collector from NYC
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      display_name:
        example: John Doe
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  dto.UpdateProfileDTO:
    properties:
      avatar_file_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      bio:
        example: Updated bio
        type: string
      display_name:
        example: John Updated
        type: string
    type: object
  dto.UserResponseDTO:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
host: localhost:4200
info:
  contact: {}
  description: REST API для управления коллекционными предметами с JWT авторизацией,
    OAuth 2.0, MinIO и Swagger.
  title: NewYear API
  version: "1.0"
paths:
...
  /auth/whoami:
    get:
      description: Возвращает профиль авторизованного пользователя. Требует валидный
        access_token в HttpOnly cookie. Нужен фронтенду для проверки статуса авторизации,
        так как HttpOnly cookies недоступны из JavaScript.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponseDTO'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - CookieAuth: []
      summary: Данные текущего пользователя
      tags:
      - Auth
...
  /profile:
    get:
      description: Возвращает профиль текущего авторизованного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfileResponseDTO'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Получить профиль пользователя
      tags:
      - Profile
    post:
      consumes:
      - application/json
      description: Обновляет профиль текущего пользователя
      parameters:
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfileResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Обновить профиль пользователя
      tags:
      - Profile
securityDefinitions:
  CookieAuth:
    description: JWT access token через HttpOnly cookie.
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"